Extracted the files for ast building into lex/building, no pipeline changes though
