+++ src/lex/parsing.rs
@@ -15,8 +15,8 @@
 //! for comprehensive documentation on using verified lex sources and AST assertions.
 
 // Parser implementations
-pub mod ast;
 pub mod builder;
+pub mod building;
 pub mod common;
 pub mod ir;
 pub mod linebased;
